--- a/be/src/vec/columns/column_nullable.h
+++ b/be/src/vec/columns/column_nullable.h
@@ -200,6 +200,12 @@ public:
 
     void insert_many_from(const IColumn& src, size_t position, size_t length) override;
 
+    void append_data_by_selector(IColumn::MutablePtr& res,
+                                 const IColumn::Selector& selector) const override;
+
+    void append_data_by_selector(IColumn::MutablePtr& res, const IColumn::Selector& selector,
+                                 size_t begin, size_t end) const override;
+
     template <typename ColumnType>
     void insert_from_with_type(const IColumn& src, size_t n) {
         const auto& src_concrete = assert_cast<const ColumnNullable&>(src);